strange compile errors

+++ MPI/CGameBoard.cpp
@@ -5,7 +5,7 @@
 #include "CGameBoard.h"
 
 //Объявляем три торы. на now-считаем, на now+next - пишем notchanged, на now+prev смотрим, что было
-CGameBoard::CGameBoard(long len, long heig):length(len), height(heig){
+CGameBoard::CGameBoard(int len, int heig):length(len), height(heig){
     now = 0;
     if (heig < 0 || len < 0)
         throw new std::exception();
@@ -26,21 +26,21 @@ void CGameBoard::setNewNow(std::vector<int>& input){
 }
 
 
-long CGameBoard::getHeight(){
+int CGameBoard::getHeight(){
     return height;
 }
 
-long CGameBoard::getLength(){
+int CGameBoard::getLength(){
     return length;
 }
 
-void CGameBoard::countCell( long index) {
+void CGameBoard::countCell( int index) {
     if (index < 0){
         index += length*height;
     }
     index = index % (length*height);
-    long x = index % length;
-    long y = index / length;
+    int x = index % length;
+    int y = index / length;
     int act  = threeToruses[(now + prev) % 3].countCell( x, y);
     threeToruses[now].action(act, threeToruses[(now + prev) % 3].getCell( x, y), x,y);
     threeToruses[(now + next) % 3].set_nothing(x, y);
@@ -63,23 +63,23 @@ CGameBoard::~CGameBoard(){
     std::cout << "Have a nice day";
 }
 
-bool CGameBoard::isCounted(long index) {
+bool CGameBoard::isCounted(int index) {
     if (index < 0){
         index += length*height;
     }
     index = index % (length*height);
-    long x = index % length;
-    long y = index / length;
+    int x = index % length;
+    int y = index / length;
     return !(threeToruses[now].getCell(x,y) == notchanged);
 }
 
-int CGameBoard::getCell(long index) {
-    long x = index % length;
-    long y = index / length;
+int CGameBoard::getCell(int index) {
+    int x = index % length;
+    int y = index / length;
     return threeToruses[now].getCell(x,y);
 }
 
-long CGameBoard::getSize() {
+int CGameBoard::getSize() {
     return length*height;
 }
